% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{djpr_connect}
\alias{djpr_connect}
\title{DJPR PostgreSQL Database Connections}
\usage{
djpr_connect(
  db = c("opendata", "official_sensitive"),
  user = c("open", "super")
)
}
\arguments{
\item{user}{}
}
\value{

}
\description{
DJPR PostgreSQL Database Connections
}
